加入容错机制

+++ app_api/views.py
@@ -1,8 +1,9 @@
 from django.http import HttpResponse
-from .models import Park, AntiPark, ParkTwo, Card, InPark
+from .models import Park, AntiPark, ParkTwo, Card, InPark, Total
 from django.views import View
 from datetime import datetime
 import time
+from django.db.models import F
 
 from .utils import dict_to_xml
 
@@ -52,108 +53,126 @@ add_time = None
 
 
 def check(request):
-    license_number = request.GET.get("license_number")
-    global license_number_card
-    global add_flag
-    global add_time
-    add_time = time.time()
-    license_number_card = license_number
-    add_flag = True
-    result = ParkTwo.objects.filter(license_number=license_number).first()
-    if result:
-        InPark.objects.create(inside=1, license_number=license_number,
-                              create_time=datetime.now())
-        return HttpResponse(1)
-    else:
-        return HttpResponse(0)
+    if request.method == 'GET':
+        license_number = request.GET.get("license_number")
+        if license_number is None:
+            return HttpResponse("error")
+        global license_number_card
+        global add_flag
+        global add_time
+        add_time = time.time()
+        license_number_card = license_number
+        add_flag = True
+        result = ParkTwo.objects.filter(license_number=license_number).first()
+        if result:
+            InPark.objects.create(inside=1, license_number=license_number,
+                                  create_time=datetime.now())
+            return HttpResponse(1)
+        else:
+            return HttpResponse(0)
 
 
 def add_car(request):
-    inside = request.POST.get("inside")
-    license_number = request.POST.get("license_number")
-    if ParkTwo.objects.filter(license_number=license_number).first() is None:
-        ParkTwo.objects.create(inside=inside, license_number=license_number,
-                               create_time=datetime.now())
-        return HttpResponse("添加成功")
-    else:
-        return HttpResponse("车牌已存在")
+    if request.method == 'POST':
+        inside = request.POST.get("inside")
+        license_number = request.POST.get("license_number")
+        if inside is None or license_number is None:
+            return HttpResponse("error")
+        if ParkTwo.objects.filter(license_number=license_number).first() is None:
+            ParkTwo.objects.create(inside=inside, license_number=license_number,
+                                   create_time=datetime.now())
+            return HttpResponse("添加成功")
+        else:
+            return HttpResponse("车牌已存在")
 
 
 def add_card(request):
-    card_number = request.POST.get("card_number")
-    if Card.objects.filter(card_number=card_number).first() is None:
-        Card.objects.create(card_number=card_number)
-        return HttpResponse("添加成功")
-    else:
-        return HttpResponse("已存在")
+    if request.method == 'POST':
+        card_number = request.POST.get("card_number")
+        if card_number is None:
+            return HttpResponse("error")
+        if Card.objects.filter(card_number=card_number).first() is None:
+            Card.objects.create(card_number=card_number)
+            return HttpResponse("添加成功")
+        else:
+            return HttpResponse("已存在")
 
 
 def delete_car(request):
-    license_number = request.GET.get("license_number")
-    if ParkTwo.objects.filter(license_number=license_number).first() is None:
-        return HttpResponse("该车牌不存在")
-    else:
-        ParkTwo.objects.filter(license_number=license_number).delete()
-        return HttpResponse("车牌删除成功！")
+    if request.method == 'GET':
+        license_number = request.GET.get("license_number")
+        if license_number is None:
+            return HttpResponse("error")
+        if ParkTwo.objects.filter(license_number=license_number).first() is None:
+            return HttpResponse("该车牌不存在")
+        else:
+            ParkTwo.objects.filter(license_number=license_number).delete()
+            return HttpResponse("车牌删除成功！")
 
 
 def delete_card(request):
-    card_number = request.GET.get("card_number")
-    if Card.objects.filter(card_number=card_number).first() is None:
-        return HttpResponse(" 该卡号不存在！")
-    else:
-        Card.objects.filter(card_number=card_number).delete()
-        return HttpResponse("卡号删除成功")
+    if request.method == 'GET':
+        card_number = request.GET.get("card_number")
+        if card_number is None:
+            return HttpResponse('error')
+        if Card.objects.filter(card_number=card_number).first() is None:
+            return HttpResponse(" 该卡号不存在！")
+        else:
+            Card.objects.filter(card_number=card_number).delete()
+            return HttpResponse("卡号删除成功")
 
 
 def check_card(request):
-    card_number = request.GET.get("card_number")
-    result = Card.objects.filter(card_number=card_number).first()
-    global add_flag
-    global license_number_card
-    if result:
-        if add_flag and not InPark.objects.filter(license_number=license_number_card).first():
-            InPark.objects.create(inside=1, license_number=license_number_card,
-                                  create_time=datetime.now())
-            add_flag = False
-            return HttpResponse(1)
+    if request.method == 'GET':
+        card_number = request.GET.get("card_number")
+        if card_number is None:
+            return HttpResponse("error")
+        result = Card.objects.filter(card_number=card_number).first()
+        global add_flag
+        global license_number_card
+        if result:
+            if add_flag and not InPark.objects.filter(license_number=license_number_card).first():
+                InPark.objects.create(inside=1, license_number=license_number_card,
+                                      create_time=datetime.now())
+                add_flag = False
+                return HttpResponse(1)
+            else:
+                license_number_card = None
+                return HttpResponse(0)
         else:
             license_number_card = None
             return HttpResponse(0)
-    else:
-        license_number_card = None
-        return HttpResponse(0)
-
-
-total = 100
 
 
 class ParkSubtract(View):
     def get(self, request):
-        global total
-        total -= 1
-        if total >= 0:
-            return HttpResponse(total)
-        else:
-            total = 0
+        if Total.objects.filter(pk=1).first().out_park < 0:
             return HttpResponse(0)
+        try:
+            Total.objects.filter(pk=1).update(out_park=F("out_park") - 1)
+            Total.objects.filter(pk=1).update(in_park=F("in_park") + 1)
+            out_park = Total.objects.filter(pk=1).first().out_park
+        except:
+            return HttpResponse(0)
+        return HttpResponse(out_park)
 
 
 class ParkAdd(View):
     def get(self, request):
-        global total
-        total += 1
-        if total <= 100:
-            return HttpResponse(total)
-        else:
-            total = 100
+        if Total.objects.filter(pk=1).first().out_park >= 100:
             return HttpResponse(100)
+        try:
+            Total.objects.filter(pk=1).update(out_park=F("out_park") + 1)
+            Total.objects.filter(pk=1).update(in_park=F("in_park") - 1)
+            out_park = Total.objects.filter(pk=1).first().out_park
+        except:
+            return HttpResponse(100)
+        return HttpResponse(out_park)
 
 
 class Pace(View):
     def get(self, request):
-        global total
-        return HttpResponse(total)
+        return HttpResponse(Total.objects.first().total)
 
 
 class ParkingNow(View):
@@ -198,6 +217,8 @@ class EnterCar(View):
 class OpenDoor(View):
     def post(self, request):
         open_door = request.POST.get("opendoor")
+        if open_door is None:
+            return HttpResponse("error")
         if open_door == "open":
             return HttpResponse(1)
         else:
@@ -205,9 +226,12 @@ class OpenDoor(View):
 
 
 def delete_car_out(request):
-    license_number = request.GET.get("license_number")
-    if InPark.objects.filter(license_number=license_number).first() is None:
-        return HttpResponse("0")
-    else:
-        InPark.objects.filter(license_number=license_number).delete()
-        return HttpResponse("1")
+    if request.method == 'GET':
+        license_number = request.GET.get("license_number")
+        if license_number is None:
+            return HttpResponse("error")
+        if InPark.objects.filter(license_number=license_number).first() is None:
+            return HttpResponse("0")
+        else:
+            InPark.objects.filter(license_number=license_number).delete()
+            return HttpResponse("1")
